refactor: move schemas to src/config/handler-config
