+++ src/handlers/close-task.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, test } from "bun:test";
 import { CoderAPIError } from "../coder-client";
 import { TestLogger } from "../infra/logger";
-import type { HandlerConfig } from "../schemas";
+import type { HandlerConfig } from "../config/handler-config";
 import {
 	MockCoderClient,
 	createMockGitHubClient,
